Stop copying input mtz files
Encountered a bug where if the two mtz files have the same file name
the pipeline won't work. Since this is quite likely and there is no
need to have a copy of the input files in the folder, it's easiest to
simply not copy the input files into the output folder.